tests/wpt/mozilla/tests/mozilla/overload_union.html

<!doctype html>
<meta charset="utf-8">
<title>Test for overload with union</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfObjectAndString({href: "a"}), "object");
  assert_equals(t.passOverloadedUnionOfObjectAndString("a"), "string");
  assert_equals(t.passOverloadedUnionOfObjectAndString(true), "boolean");

}, "Overload of union of object and string works")

test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfObjectAndNumber({href: "a"}), "object");
  assert_equals(t.passOverloadedUnionOfObjectAndNumber(100), "number");
  assert_equals(t.passOverloadedUnionOfObjectAndNumber(true), "boolean");

}, "Overload of union of object and number works")

test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfObjectAndBoolean({href: "a"}), "object");
  assert_equals(t.passOverloadedUnionOfObjectAndBoolean(true), "boolean");
  assert_equals(t.passOverloadedUnionOfObjectAndBoolean(100), "number");

}, "Overload of union of object and boolean works")

test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfStringAndNumber("a"), "string");
  assert_equals(t.passOverloadedUnionOfStringAndNumber(100), "number");
  assert_equals(t.passOverloadedUnionOfStringAndNumber(true), "boolean");

}, "Overload of union of string and number works")

test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfStringAndBoolean("a"), "string");
  assert_equals(t.passOverloadedUnionOfStringAndBoolean(true), "boolean");
  assert_equals(t.passOverloadedUnionOfStringAndBoolean(100), "number");

}, "Overload of union of string and boolean works")

test(function() {
  var t = new TestBinding();

  assert_equals(t.passOverloadedUnionOfNumberAndBoolean(100), "number");
  assert_equals(t.passOverloadedUnionOfNumberAndBoolean(true), "boolean");
  assert_equals(t.passOverloadedUnionOfNumberAndBoolean("a"), "string");

}, "Overload of union of number and boolean works")
</script>
